# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Store API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/cms/books": {
            "get": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Danh sách Books ",
                "operationId": "cms_book_list",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/books/store": {
            "post": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Thêm book",
                "operationId": "cms_book_create",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dữ liệu sách cần thêm",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tên book"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publisher_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2002-09-29"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/books/{id}": {
            "get": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Chi tiết thông tin của Book",
                "operationId": "cms_book_detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của book",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/books/update/{id}": {
            "put": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Sửa book",
                "operationId": "cms_book_update",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id của book",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dữ liệu",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tên book"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publisher_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2002-09-29"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/books/soft-delete": {
            "delete": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Xoá mềm nhiều Book",
                "operationId": "cms_book_softDelete",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Danh sách ID cần xoá",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Xoá mềm sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Soft delete success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Thiếu danh sách ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No IDs provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/books/trash": {
            "get": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Danh sách Books đã xoá",
                "operationId": "cms_book_trash",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/books/recovery": {
            "post": {
                "tags": [
                    "CMS Books"
                ],
                "summary": "Phục hồi book",
                "operationId": "cms_book_recovery",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Danh sách ID cần phục hồi",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}